backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
django-filter==23.5
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-storages==1.14.2
boto3==1.34.7
gunicorn==21.2.0